Best way to use PlayerClass and enums!!!
