--- src/main/java/lar/minecraft/hg/PlayerAction.java
@@ -1,8 +1,0 @@
-package lar.minecraft.hg;
-
-import org.bukkit.entity.Player;
-
-@FunctionalInterface
-public interface PlayerAction {
-    void perform(Player player);
-}
