--- src/main/java/lar/minecraft/hg/enums/PlayerClassEnum.java
@@ -1,8 +1,0 @@
-package lar.minecraft.hg.enums;
-
-public enum PlayerClassEnum {
-	BOWMAN,
-	ARMORED,
-	DOGLOVER,
-	LAVAMAN
-}
